Get shotgun hillclimb working well. Ish.
It's highly motivated to drain the battery at the end of the day, as
that's how it gets the most profits. THe behaviour is therefore highly
dependent on the price that you attach to the remaining battery charge.

It also has a habit of charging from solar late in the day, rather than
early.